+++ modules/School_Setup/includes/Modules.inc.php
@@ -296,14 +296,17 @@ if ( ! $_REQUEST['modfunc'] )
 		&& is_writable( 'modules/' ) )
 	{
 		// @since 6.4 Add-on zip upload.
-		echo '<form action="' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&tab=' . $_REQUEST['tab'] .
+		echo '<br /><form action="' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&tab=' . $_REQUEST['tab'] .
 			'&modfunc=upload' ) . '" method="POST" enctype="multipart/form-data">';
 
-		echo button( 'add' ) . ' ';
-
-		echo FileInput( 'upload', '', 'required accept=".zip"', FileUploadMaxSize() );
+		echo FileInput( 'upload', '', 'required accept=".zip"' );
 
 		echo SubmitButton( _( 'Upload' ) );
+
+		echo FormatInputTitle(
+			button( 'add', '', '', 'smaller' ) . ' ' . _( 'Module' ) . ' (.zip)',
+			'upload'
+		);
 
 		echo '</form>';
 	}